# docs/swagger.yaml
basePath: /api/v1/
definitions:
  controller._ResponseCommunityDetailList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 业务相应码
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.CommunityDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponseCommunityList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 业务相应码
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.Community'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponsePostList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 业务相应码
      data:
        description: 数据
        items:
          $ref: '#/definitions/models.ApiPostDetail'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponseUserLogin:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 业务相应码
      data:
        allOf:
        - $ref: '#/definitions/models.UserLogin'
        description: 数据
      message:
        description: 提示信息
        type: string
    type: object
  models.ApiPostDetail:
    properties:
      author_name:
        description: 帖子作者用户名
        type: string
      community:
        allOf:
        - $ref: '#/definitions/models.CommunityDetail'
        description: 嵌入社区信息
      post:
        allOf:
        - $ref: '#/definitions/models.Post'
        description: 嵌入帖子信息
      vote_agree_num:
        description: 投票赞成数
        type: integer
    type: object
  models.Community:
    properties:
      community_id:
        description: 社区id
        type: integer
      community_name:
        description: 社区名称
        type: string
    type: object
  models.CommunityDetail:
    properties:
      community_id:
        description: 社区id
        type: integer
      community_name:
        description: 社区名称
        type: string
      create_time:
        description: 社区创建时间
        type: string
      introduction:
        description: 社区介绍
        type: string
      update_time:
        description: 社区更新时间
        type: string
    type: object
  models.ParamLogin:
    properties:
      password:
        description: 登录人密码
        type: string
      username:
        description: 登录人姓名
        type: string
    required:
    - password
    - username
    type: object
  models.ParamPostCreate:
    properties:
      community_id:
        description: 社区id
        type: integer
      content:
        description: 帖子内容
        type: string
      title:
        description: 帖子标题
        type: string
    required:
    - community_id
    - content
    - title
    type: object
  models.ParamSignUp:
    properties:
      age:
        description: 注册人年龄
        maximum: 130
        minimum: 1
        type: integer
      email:
        description: 注册人邮箱
        type: string
      gender:
        description: 注册人性别
        enum:
        - 男
        - 女
        type: string
      password:
        description: 注册人密码
        type: string
      re_password:
        description: 注册人重复密码
        type: string
      username:
        description: 注册人用户名
        type: string
    required:
    - email
    - gender
    - password
    - re_password
    - username
    type: object
  models.ParamVoteData:
    properties:
      kind:
        description: 投票帖子赞成(1)\反对(-1)\取消投票(0)
        enum:
        - 1
        - 0
        - -1
        example: "0"
        type: string
      post_id:
        description: UserID 从请求中获取当前用户
        example: "0"
        type: string
    required:
    - post_id
    type: object
  models.Post:
    properties:
      author_id:
        description: 帖子作者id
        example: "0"
        type: string
      community_id:
        description: 社区id
        type: integer
      content:
        description: 帖子内容
        type: string
      create_time:
        description: 帖子创建时间
        type: string
      post_id:
        description: 帖子id
        example: "0"
        type: string
      status:
        description: 帖子状态
        type: integer
      title:
        description: 帖子标题
        type: string
      update_time:
        description: 帖子更新时间
        type: string
    required:
    - community_id
    - content
    - title
    type: object
  models.UserLogin:
    properties:
      token:
        description: access token
        type: string
      user_id:
        description: 用户id
        type: integer
      user_name:
        description: 用户名
        type: string
    type: object
  response.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
host: 127.0.0.1:8081
info:
  contact:
    email: none
    name: lucien
    url: none
  description: This is a go-web project from https://www.bilibili.com/cheese/play/ep265306
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: bluebell go-web
  version: "1.0"
paths:
  /community:
    get:
      consumes:
      - application/json
      description: 查询社区信息返回列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunityList'
      summary: 查询社区信息接口
      tags:
      - 社区接口
  /community/{community_id}:
    get:
      consumes:
      - application/json
      description: 查询社区分类详情信息返回列表
      parameters:
      - description: 社区id
        example: 2
        in: path
        name: community_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunityDetailList'
      summary: 查询社区分类详情信息接口
      tags:
      - 社区接口
  /login:
    post:
      consumes:
      - application/json
      description: 根据所填用户名和密码进行登录
      parameters:
      - description: 用户登录
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseUserLogin'
      summary: 用户登录接口
      tags:
      - 用户接口
  /post:
    post:
      consumes:
      - application/json
      description: 发布帖子
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 请求参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/models.ParamPostCreate'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 发布帖子
      tags:
      - 帖子接口
  /post/{post_id}:
    get:
      consumes:
      - application/json
      description: 根据帖子id获取帖子分类详情接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 帖子id
        example: "0"
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子分类详情接口
      tags:
      - 帖子接口
  /posts:
    get:
      consumes:
      - application/json
      description: 根据分页参数获取帖子列表分页展示
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 获取帖子列表的页码
        example: 1
        in: query
        name: page_num
        type: integer
      - description: 获取帖子列表的数量
        example: 5
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      summary: 获取帖子列表分页展示接口
      tags:
      - 帖子接口
  /searchposts:
    get:
      consumes:
      - application/json
      description: 可按社区按时间或分数排序查询帖子列表接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 社区id可以为空
        in: query
        name: community_id
        type: integer
      - description: 排序方式 按照时间或者投票分数
        example: score
        in: query
        name: order
        type: string
      - description: 获取帖子列表的页码
        example: 1
        in: query
        name: page_num
        type: integer
      - description: 获取帖子列表的数量
        example: 5
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePostList'
      security:
      - ApiKeyAuth: []
      summary: 升级版帖子列表接口
      tags:
      - 帖子接口
  /signup:
    post:
      consumes:
      - application/json
      description: 根据用户所填信息进行注册
      parameters:
      - description: 用户注册
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamSignUp'
      produces:
      - application/json
      responses: {}
      summary: 用户注册接口
      tags:
      - 用户接口
  /vote:
    post:
      consumes:
      - application/json
      description: 根据帖子id进行投票
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 用户投票
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamVoteData'
      produces:
      - application/json
      responses: {}
      summary: 用户投票接口
      tags:
      - 用户接口
swagger: "2.0"